Move code from FPSAssetLoaderDelegate to FPSLicenseHelper
