+++ Classes/FairPlayDRM/FPSLicenseHelper.swift
@@ -22,12 +22,14 @@ enum FairPlayError : Error {
 class FPSLicenseHelper {
     
     let assetId: String
-    let appCertificate: Data
-    let licenseUrl: URL
-    let forceDownload: Bool
     
+    let appCertificate: Data?
+    let licenseUrl: URL?
+    
+    let forceDownload: Bool
+    let shouldPersist: Bool
 
-    init(assetId: String, params: FairPlayDRMParams, forceDownload: Bool = false) throws {
+    init(assetId: String, params: FairPlayDRMParams, shouldPersist: Bool, forceDownload: Bool = false) throws {
         self.assetId = assetId
         
         guard let cert = params.fpsCertificate else {throw FairPlayError.noAppCertificate}
@@ -37,12 +39,22 @@ class FPSLicenseHelper {
         self.licenseUrl = url
         
         self.forceDownload = forceDownload
+        self.shouldPersist = shouldPersist
     }
     
-    @available(iOS 10.3, *)
-    convenience init(keyIdentifier: Any?, params: FairPlayDRMParams) throws {
-        try self.init(assetId: FPSLicenseHelper.getAssetId(keyIdentifier: keyIdentifier), params: params)
-    }
+    init(assetId: String) {
+        self.assetId = assetId
+        self.forceDownload = false
+        self.appCertificate = nil
+        self.licenseUrl = nil
+        self.shouldPersist = true
+   }
+    
+//    @available(iOS 10.3, *)
+//    convenience init(keyIdentifier: Any?, params: FairPlayDRMParams) throws {
+//        self.ini
+//        try self.init(assetId: FPSLicenseHelper.getAssetId(keyIdentifier: keyIdentifier), params: params, shouldPersist: )
+//    }
     
     @available(iOS 10.3, *)
     func requestPersistableContentKeys() {
@@ -65,7 +77,9 @@ class FPSLicenseHelper {
     
     func performCKCRequest(_ spcData: Data) throws -> Data {
         
-        var request = URLRequest(url: licenseUrl)
+        guard let url = self.licenseUrl else { throw FairPlayError.noLicenseURL }
+        
+        var request = URLRequest(url: url)
         request.setValue("application/octet-stream", forHTTPHeaderField: "Content-Type")
         request.httpBody = spcData.base64EncodedData()
         request.httpMethod = "POST"
@@ -101,10 +115,118 @@ class FPSLicenseHelper {
         return ckc
     }
 
+    func fetchLicense(resourceLoadingRequest: AVAssetResourceLoadingRequest, usePersistence: Bool, done: (Error?)->Void) throws {
+        // Check if we have an existing key on disk for this asset.
+        let keyLocation = FairPlayUtils.urlForPersistableContentKey(withContentKeyIdentifier: assetId)
+        
+        if !forceDownload && FileManager.default.fileExists(atPath: keyLocation.path) {
+            guard let dataRequest = resourceLoadingRequest.dataRequest else {
+                PKLog.error("Error loading contents of content key file.")
+                let error = NSError(domain: FPSAssetLoaderDelegate.errorDomain, code: -2, userInfo: nil)
+                resourceLoadingRequest.finishLoading(with: error)
+                done(error)
+                return
+            }
+            
+            if let storedKey = try? Data.init(contentsOf: keyLocation) {
+                
+                // Pass the persistedContentKeyData into the dataRequest to complete the content key request.
+                dataRequest.respond(with: storedKey)
+                resourceLoadingRequest.finishLoading()
+                done(nil)
+            }
+            return
+        }
+
+        guard let appCert = self.appCertificate else { throw FairPlayError.noAppCertificate }
+        
+        var resourceLoadingRequestOptions: [String: AnyObject]? = nil
+        
+        if #available(iOS 10.0, *), usePersistence {
+            resourceLoadingRequestOptions = [AVAssetResourceLoadingRequestStreamingContentKeyRequestRequiresPersistentKey: true as AnyObject]
+        }
+        
+        let spcData: Data!
+        
+        do {
+            /*
+             To obtain the Server Playback Context (SPC), we call
+             AVAssetResourceLoadingRequest.streamingContentKeyRequestData(forApp:contentIdentifier:options:)
+             using the information we obtained earlier.
+             */
+            spcData = try resourceLoadingRequest.streamingContentKeyRequestData(forApp: appCert, contentIdentifier: assetId.data(using: .utf8)!, options: resourceLoadingRequestOptions)
+            PKLog.debug("Got spcData with", spcData.count, "bytes")
+        } catch let error as NSError {
+            PKLog.error("Error obtaining key request data: \(error.domain) reason: \(String(describing: error.localizedFailureReason))")
+            resourceLoadingRequest.finishLoading(with: error)
+            done(error)
+            return
+        }
+        
+        do {
+            let ckcData = try performCKCRequest(spcData)
+            self.handleCKCData(resourceLoadingRequest, ckcData: ckcData, done: done)
+        } catch {
+            PKLog.error("Error occured while loading FairPlay license:", error)
+        }
+    }
+
+    func handleCKCData(_ resourceLoadingRequest: AVAssetResourceLoadingRequest, ckcData: Data, done: (Error?)->Void) {
+        
+        // Check if this reuqest is the result of a potential AVAssetDownloadTask.
+        if #available(iOS 10.0, *), shouldPersist {
+            /* Since this request is the result of an AVAssetDownloadTask, we should get the secure persistent content key.
+             Obtain a persistable content key from a context.
+             
+             The data returned from this method may be used to immediately satisfy an
+             AVAssetResourceLoadingDataRequest, as well as any subsequent requests for the same key url.
+             
+             The value of AVAssetResourceLoadingContentInformationRequest.contentType must be set to AVStreamingKeyDeliveryPersistentContentKeyType when responding with data created with this method.
+             */
+            do {
+                let persistentContentKeyData = try resourceLoadingRequest.persistentContentKey(fromKeyVendorResponse: ckcData, options: nil)
+                
+                // Save the persistentContentKeyData onto disk for use in the future.
+                PKLog.debug("Saving persistentContentKeyData")
+                try FairPlayUtils.writePersistableContentKey(contentKey: persistentContentKeyData, withContentKeyIdentifier: assetId)
+                
+                guard let dataRequest = resourceLoadingRequest.dataRequest else {
+                    PKLog.error("no data is being requested in loadingRequest")
+                    let error = NSError(domain: FPSAssetLoaderDelegate.errorDomain, code: -6, userInfo: nil)
+                    resourceLoadingRequest.finishLoading(with: error)
+                    done(error)
+                    return
+                }
+                // Provide data to the loading request.
+                dataRequest.respond(with: persistentContentKeyData)
+                resourceLoadingRequest.finishLoading()  // Treat the processing of the request as complete.
+                done(nil)
+            } catch {
+                PKLog.error("Error creating persistent content key: \(error)")
+                resourceLoadingRequest.finishLoading(with: error)
+                done(error)
+                return
+            }
+        } else {
+            
+            guard let dataRequest = resourceLoadingRequest.dataRequest else {
+                PKLog.error("no data is being requested in loadingRequest")
+                let error = NSError(domain: FPSAssetLoaderDelegate.errorDomain, code: -6, userInfo: nil)
+                resourceLoadingRequest.finishLoading(with: error)
+                done(error)
+                return
+            }
+            
+            // Provide data to the loading request.
+            dataRequest.respond(with: ckcData)
+            resourceLoadingRequest.finishLoading()  // Treat the processing of the request as complete.
+            done(nil)
+        }
+    }
     
     @available(iOS 10.3, *)
-    func fetchLicense(with keyRequest: AVContentKeyRequest, done: @escaping ()->Void) throws {
-        let assetId = try FPSLicenseHelper.getAssetId(keyRequest)
+    func fetchLicense(contentKeyRequest: AVContentKeyRequest, done: @escaping ()->Void) throws {
+        let assetId = self.assetId
         let keyLocation = FairPlayUtils.urlForPersistableContentKey(withContentKeyIdentifier: assetId)
         
         if !forceDownload && FileManager.default.fileExists(atPath: keyLocation.path) {
@@ -114,18 +236,21 @@ class FPSLicenseHelper {
                 let keyResponse = AVContentKeyResponse(fairPlayStreamingKeyResponseData: storedKey)
                 
                 // Provide the content key response to make protected content available for processing.
-                keyRequest.processContentKeyResponse(keyResponse)
+                contentKeyRequest.processContentKeyResponse(keyResponse)
             }
             return
         }
         
-        keyRequest.makeStreamingContentKeyRequestData(forApp: appCertificate, 
+        guard let appCert = self.appCertificate else { throw FairPlayError.noAppCertificate }
+        
+        contentKeyRequest.makeStreamingContentKeyRequestData(forApp: appCert, 
                                                       contentIdentifier: assetId.data(using: .utf8)!, 
                                                       options: [AVContentKeyRequestProtocolVersionsKey: [1]]) { [weak self] (spcData, error) in
             
+                                                        
             guard let strongSelf = self else { return }
             if let error = error {
-                keyRequest.processContentKeyResponseError(error)
+                contentKeyRequest.processContentKeyResponseError(error)
                 done()
                 //                    strongSelf.pendingPersistableContentKeyIdentifiers.remove(assetId)
                 return
@@ -139,7 +264,7 @@ class FPSLicenseHelper {
                 let ckcData = try strongSelf.performCKCRequest(spcData)
                 var keyData: Data
                 
-                if let persReq = keyRequest as? AVPersistableContentKeyRequest {
+                if let persReq = contentKeyRequest as? AVPersistableContentKeyRequest {
                     let pKey = try persReq.persistableContentKey(fromKeyVendorResponse: ckcData, options: nil)
                     keyData = pKey
                     
@@ -157,12 +282,12 @@ class FPSLicenseHelper {
                 /*
                  Provide the content key response to make protected content available for processing.
                  */
-                keyRequest.processContentKeyResponse(keyResponse)
+                contentKeyRequest.processContentKeyResponse(keyResponse)
                 
                 done()
                 //                    strongSelf.pendingPersistableContentKeyIdentifiers.remove(assetId)
             } catch {
-                keyRequest.processContentKeyResponseError(error)
+                contentKeyRequest.processContentKeyResponseError(error)
                 
                 done()
                 //                    strongSelf.pendingPersistableContentKeyIdentifiers.remove(assetId)
